+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### v2.0
+
+- settings activity
+- start with first bookmark or …
+- … with bookmark list
+- editable bookmarks
+- ui improvements
+- splashscreen
+
 ### v1.2
 
 - added english